Trying to minimize use of ToyVirtualWorld
Changed project name string -> ‘VirtualWorldColorConstancy’

Removed level of indirection on some output files, so that the data just go into ‘VirtualWorldColorConstancy’ rather than ‘VirtualWorldColorConstancy/ToyVirtualWorld’.  Possible this is a mistake, we will find out.

There are some routines with ToyVirtualWorld in their names which we should also change, but we did not do that yet.

Signed-off-by: David Brainard <brainard@psych.upenn.edu>