change ill Reflec naming
The naming convetion was changes to precede zeros before the numbers to make sure that while reading the files the right file is read. If the files are named illuminant1.spd illuminant2.spd and illuminant10.spd, then the second file is illuminant10.spd not illuminant2.spd because 1 has preference over 2.